app/requirements.txt
Flask==3.0.1
Flask_Cors==4.0.0
Pillow==10.2.0
Gunicorn==21.2.0
